Reworked Wedding Ceremony Preparation
Added preparation files for Wedding Ceremonies Rework. This includes the debug/helper panel, the actual wedding script (currently works with Murder special action). Updates f.lua to add GetAlignmentValue. ms_debug_weddings.lua has the premisces of the next part that is being worked on: ability to add options to increase the impressiveness of the ceremony.