Lodge Crash
Temporarily deactivated parts of the code which seems to cause a game crash. The investigation continues, there is no proper way to know why this is happening. However with this code deactivated, I believe the culprit is the stored Count.Guests and Count.CanOrder which somehow break the CleanUp(). I will continue on tomorrow when waking up.